circuits/ecc/foreign: reorganize gates in two folders
