+++ circuits/src/ecc/foreign/gates/edwards/mod.rs
@@ -11,5 +11,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub(crate) mod edwards;
-pub(crate) mod weierstrass;
+pub(crate) mod coord;
