+++ circuits/src/ecc/foreign/gates/weierstrass/lambda_squared.rs
